+++ packages/realm_dart/lib/src/native/realm_bindings.dart
@@ -3343,6 +3343,20 @@ class RealmLibrary {
               ffi.Pointer<realm_thread_safe_reference_t>,
               ffi.Pointer<realm_async_error_t>)>();
 
+  void realm_dart_attach_logger(
+    int port,
+  ) {
+    return _realm_dart_attach_logger(
+      port,
+    );
+  }
+
+  late final _realm_dart_attach_loggerPtr =
+      _lookup<ffi.NativeFunction<ffi.Void Function(Dart_Port)>>(
+          'realm_dart_attach_logger');
+  late final _realm_dart_attach_logger =
+      _realm_dart_attach_loggerPtr.asFunction<void Function(int)>();
+
   ffi.Pointer<realm_scheduler_t> realm_dart_create_scheduler(
     int isolateId,
     int port,
@@ -3636,6 +3650,20 @@ class RealmLibrary {
       _realm_dart_delete_persistent_handlePtr
           .asFunction<void Function(ffi.Pointer<ffi.Void>)>();
 
+  void realm_dart_detach_logger(
+    int port,
+  ) {
+    return _realm_dart_detach_logger(
+      port,
+    );
+  }
+
+  late final _realm_dart_detach_loggerPtr =
+      _lookup<ffi.NativeFunction<ffi.Void Function(Dart_Port)>>(
+          'realm_dart_detach_logger');
+  late final _realm_dart_detach_logger =
+      _realm_dart_detach_loggerPtr.asFunction<void Function(int)>();
+
   /// implemented for Android only
   ffi.Pointer<ffi.Char> realm_dart_get_bundle_id() {
     return _realm_dart_get_bundle_id();
@@ -3709,20 +3737,15 @@ class RealmLibrary {
           void Function(ffi.Pointer<ffi.Void>, realm_http_request_t,
               ffi.Pointer<ffi.Void>)>();
 
-  /// Returns `true` if Realm Core logger was initialized.
-  bool realm_dart_init_core_logger(
-    int level,
-  ) {
-    return _realm_dart_init_core_logger(
-      level,
-    );
+  void realm_dart_init_debug_logger() {
+    return _realm_dart_init_debug_logger();
   }
 
-  late final _realm_dart_init_core_loggerPtr =
-      _lookup<ffi.NativeFunction<ffi.Bool Function(ffi.Int32)>>(
-          'realm_dart_init_core_logger');
-  late final _realm_dart_init_core_logger =
-      _realm_dart_init_core_loggerPtr.asFunction<bool Function(int)>();
+  late final _realm_dart_init_debug_loggerPtr =
+      _lookup<ffi.NativeFunction<ffi.Void Function()>>(
+          'realm_dart_init_debug_logger');
+  late final _realm_dart_init_debug_logger =
+      _realm_dart_init_debug_loggerPtr.asFunction<void Function()>();
 
   void realm_dart_initializeDartApiDL(
     ffi.Pointer<ffi.Void> data,
@@ -3767,23 +3790,24 @@ class RealmLibrary {
   late final _realm_dart_library_version = _realm_dart_library_versionPtr
       .asFunction<ffi.Pointer<ffi.Char> Function()>();
 
-  void realm_dart_log_message_for_testing(
+  void realm_dart_log(
     int level,
+    ffi.Pointer<ffi.Char> category,
     ffi.Pointer<ffi.Char> message,
   ) {
-    return _realm_dart_log_message_for_testing(
+    return _realm_dart_log(
       level,
+      category,
       message,
     );
   }
 
-  late final _realm_dart_log_message_for_testingPtr = _lookup<
-          ffi
-          .NativeFunction<ffi.Void Function(ffi.Int32, ffi.Pointer<ffi.Char>)>>(
-      'realm_dart_log_message_for_testing');
-  late final _realm_dart_log_message_for_testing =
-      _realm_dart_log_message_for_testingPtr
-          .asFunction<void Function(int, ffi.Pointer<ffi.Char>)>();
+  late final _realm_dart_logPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Void Function(ffi.Int32, ffi.Pointer<ffi.Char>,
+              ffi.Pointer<ffi.Char>)>>('realm_dart_log');
+  late final _realm_dart_log = _realm_dart_logPtr.asFunction<
+      void Function(int, ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>)>();
 
   ffi.Pointer<ffi.Void> realm_dart_object_to_persistent_handle(
     Object handle,
@@ -3814,20 +3838,6 @@ class RealmLibrary {
   late final _realm_dart_persistent_handle_to_object =
       _realm_dart_persistent_handle_to_objectPtr
           .asFunction<Object Function(ffi.Pointer<ffi.Void>)>();
-
-  void realm_dart_release_logger(
-    int port,
-  ) {
-    return _realm_dart_release_logger(
-      port,
-    );
-  }
-
-  late final _realm_dart_release_loggerPtr =
-      _lookup<ffi.NativeFunction<ffi.Void Function(Dart_Port)>>(
-          'realm_dart_release_logger');
-  late final _realm_dart_release_logger =
-      _realm_dart_release_loggerPtr.asFunction<void Function(int)>();
 
   void realm_dart_return_string_callback(
     ffi.Pointer<ffi.Void> userdata,
@@ -3867,22 +3877,6 @@ class RealmLibrary {
               ffi.Pointer<ffi.Void>)>>('realm_dart_scheduler_invoke');
   late final _realm_dart_scheduler_invoke = _realm_dart_scheduler_invokePtr
       .asFunction<void Function(int, ffi.Pointer<ffi.Void>)>();
-
-  void realm_dart_set_log_level(
-    int level,
-    int port,
-  ) {
-    return _realm_dart_set_log_level(
-      level,
-      port,
-    );
-  }
-
-  late final _realm_dart_set_log_levelPtr =
-      _lookup<ffi.NativeFunction<ffi.Void Function(ffi.Int32, Dart_Port)>>(
-          'realm_dart_set_log_level');
-  late final _realm_dart_set_log_level =
-      _realm_dart_set_log_levelPtr.asFunction<void Function(int, int)>();
 
   bool realm_dart_sync_after_reset_handler_callback(
     ffi.Pointer<ffi.Void> userdata,
@@ -3927,26 +3921,6 @@ class RealmLibrary {
   late final _realm_dart_sync_before_reset_handler_callback =
       _realm_dart_sync_before_reset_handler_callbackPtr.asFunction<
           bool Function(ffi.Pointer<ffi.Void>, ffi.Pointer<realm_t>)>();
-
-  void realm_dart_sync_client_log_callback(
-    ffi.Pointer<ffi.Void> userdata,
-    int level,
-    ffi.Pointer<ffi.Char> message,
-  ) {
-    return _realm_dart_sync_client_log_callback(
-      userdata,
-      level,
-      message,
-    );
-  }
-
-  late final _realm_dart_sync_client_log_callbackPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Void Function(ffi.Pointer<ffi.Void>, ffi.Int32,
-              ffi.Pointer<ffi.Char>)>>('realm_dart_sync_client_log_callback');
-  late final _realm_dart_sync_client_log_callback =
-      _realm_dart_sync_client_log_callbackPtr.asFunction<
-          void Function(ffi.Pointer<ffi.Void>, int, ffi.Pointer<ffi.Char>)>();
 
   void realm_dart_sync_connection_state_changed_callback(
     ffi.Pointer<ffi.Void> userdata,
@@ -9061,13 +9035,11 @@ class RealmLibrary {
   /// Install the default logger
   void realm_set_log_callback(
     realm_log_func_t arg0,
-    int arg1,
     ffi.Pointer<ffi.Void> userdata,
     realm_free_userdata_func_t userdata_free,
   ) {
     return _realm_set_log_callback(
       arg0,
-      arg1,
       userdata,
       userdata_free,
     );
@@ -9075,10 +9047,10 @@ class RealmLibrary {
 
   late final _realm_set_log_callbackPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Void Function(realm_log_func_t, ffi.Int32, ffi.Pointer<ffi.Void>,
+          ffi.Void Function(realm_log_func_t, ffi.Pointer<ffi.Void>,
               realm_free_userdata_func_t)>>('realm_set_log_callback');
   late final _realm_set_log_callback = _realm_set_log_callbackPtr.asFunction<
-      void Function(realm_log_func_t, int, ffi.Pointer<ffi.Void>,
+      void Function(realm_log_func_t, ffi.Pointer<ffi.Void>,
           realm_free_userdata_func_t)>();
 
   void realm_set_log_level(
@@ -11692,6 +11664,8 @@ class _SymbolAddresses {
                   ffi.Pointer<realm_async_error_t>)>>
       get realm_dart_async_open_task_callback =>
           _library._realm_dart_async_open_task_callbackPtr;
+  ffi.Pointer<ffi.NativeFunction<ffi.Void Function(Dart_Port)>>
+      get realm_dart_attach_logger => _library._realm_dart_attach_loggerPtr;
   ffi.Pointer<
           ffi.NativeFunction<
               ffi.Pointer<realm_scheduler_t> Function(ffi.Uint64, Dart_Port)>>
@@ -11769,6 +11743,8 @@ class _SymbolAddresses {
   ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
       get realm_dart_delete_persistent_handle =>
           _library._realm_dart_delete_persistent_handlePtr;
+  ffi.Pointer<ffi.NativeFunction<ffi.Void Function(Dart_Port)>>
+      get realm_dart_detach_logger => _library._realm_dart_detach_loggerPtr;
   ffi.Pointer<ffi.NativeFunction<ffi.Pointer<ffi.Char> Function()>>
       get realm_dart_get_bundle_id => _library._realm_dart_get_bundle_idPtr;
   ffi.Pointer<ffi.NativeFunction<ffi.Pointer<ffi.Char> Function()>>
@@ -11785,9 +11761,9 @@ class _SymbolAddresses {
           ffi.Void Function(ffi.Pointer<ffi.Void>, realm_http_request_t,
               ffi.Pointer<ffi.Void>)>> get realm_dart_http_request_callback =>
       _library._realm_dart_http_request_callbackPtr;
-  ffi.Pointer<ffi.NativeFunction<ffi.Bool Function(ffi.Int32)>>
-      get realm_dart_init_core_logger =>
-          _library._realm_dart_init_core_loggerPtr;
+  ffi.Pointer<ffi.NativeFunction<ffi.Void Function()>>
+      get realm_dart_init_debug_logger =>
+          _library._realm_dart_init_debug_loggerPtr;
   ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
       get realm_dart_initializeDartApiDL =>
           _library._realm_dart_initializeDartApiDLPtr;
@@ -11799,18 +11775,16 @@ class _SymbolAddresses {
   ffi.Pointer<ffi.NativeFunction<ffi.Pointer<ffi.Char> Function()>>
       get realm_dart_library_version => _library._realm_dart_library_versionPtr;
   ffi.Pointer<
-          ffi
-          .NativeFunction<ffi.Void Function(ffi.Int32, ffi.Pointer<ffi.Char>)>>
-      get realm_dart_log_message_for_testing =>
-          _library._realm_dart_log_message_for_testingPtr;
+          ffi.NativeFunction<
+              ffi.Void Function(
+                  ffi.Int32, ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>)>>
+      get realm_dart_log => _library._realm_dart_logPtr;
   ffi.Pointer<ffi.NativeFunction<ffi.Pointer<ffi.Void> Function(ffi.Handle)>>
       get realm_dart_object_to_persistent_handle =>
           _library._realm_dart_object_to_persistent_handlePtr;
   ffi.Pointer<ffi.NativeFunction<ffi.Handle Function(ffi.Pointer<ffi.Void>)>>
       get realm_dart_persistent_handle_to_object =>
           _library._realm_dart_persistent_handle_to_objectPtr;
-  ffi.Pointer<ffi.NativeFunction<ffi.Void Function(Dart_Port)>>
-      get realm_dart_release_logger => _library._realm_dart_release_loggerPtr;
   ffi.Pointer<
           ffi.NativeFunction<
               ffi.Void Function(ffi.Pointer<ffi.Void>, ffi.Pointer<ffi.Char>,
@@ -11822,8 +11796,6 @@ class _SymbolAddresses {
           .NativeFunction<ffi.Void Function(ffi.Uint64, ffi.Pointer<ffi.Void>)>>
       get realm_dart_scheduler_invoke =>
           _library._realm_dart_scheduler_invokePtr;
-  ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Int32, Dart_Port)>>
-      get realm_dart_set_log_level => _library._realm_dart_set_log_levelPtr;
   ffi.Pointer<
       ffi.NativeFunction<
           ffi.Bool Function(
@@ -11837,12 +11809,6 @@ class _SymbolAddresses {
               ffi.Bool Function(ffi.Pointer<ffi.Void>, ffi.Pointer<realm_t>)>>
       get realm_dart_sync_before_reset_handler_callback =>
           _library._realm_dart_sync_before_reset_handler_callbackPtr;
-  ffi.Pointer<
-          ffi.NativeFunction<
-              ffi.Void Function(
-                  ffi.Pointer<ffi.Void>, ffi.Int32, ffi.Pointer<ffi.Char>)>>
-      get realm_dart_sync_client_log_callback =>
-          _library._realm_dart_sync_client_log_callbackPtr;
   ffi.Pointer<
           ffi.NativeFunction<
               ffi.Void Function(ffi.Pointer<ffi.Void>, ffi.Int32, ffi.Int32)>>
@@ -12517,10 +12483,14 @@ typedef realm_log_func_t
     = ffi.Pointer<ffi.NativeFunction<realm_log_func_tFunction>>;
 typedef realm_log_func_tFunction = ffi.Void Function(
     ffi.Pointer<ffi.Void> userdata,
+    ffi.Pointer<ffi.Char> category,
     ffi.Int32 level,
     ffi.Pointer<ffi.Char> message);
 typedef Dartrealm_log_func_tFunction = void Function(
-    ffi.Pointer<ffi.Void> userdata, int level, ffi.Pointer<ffi.Char> message);
+    ffi.Pointer<ffi.Void> userdata,
+    ffi.Pointer<ffi.Char> category,
+    int level,
+    ffi.Pointer<ffi.Char> message);
 
 /// Logging */
 /// // equivalent to realm::util::Logger::Level in util/logger.hpp and must be kept in sync.